Fix controllers authorization

+++ protected/controllers/TvobjectController.php
@@ -14,34 +14,7 @@ class TvobjectController extends Controller
 	public function filters()
 	{
 		return array(
-			'accessControl', // perform access control for CRUD operations
 			'postOnly + delete', // we only allow deletion via POST request
-		);
-	}
-
-	/**
-	 * Specifies the access control rules.
-	 * This method is used by the 'accessControl' filter.
-	 * @return array access control rules
-	 */
-	public function accessRules()
-	{
-		return array(
-			array('allow',  // allow all users to perform 'index' and 'view' actions
-				'actions'=>array('index','view'),
-				'users'=>array('*'),
-			),
-			array('allow', // allow teacher user to perform 'create' and 'update' actions
-				'actions'=>array('create','update'),
-				'roles'=>array('teacher'),
-			),
-			array('allow', // allow admin user to perform 'admin' and 'delete' actions
-				'actions'=>array('admin','delete'),
-				'roles'=>array('admin'),
-			),
-			array('deny',  // deny all users
-				'users'=>array('*'),
-			),
 		);
 	}
 
@@ -51,9 +24,16 @@ class TvobjectController extends Controller
 	 */
 	public function actionView($id)
 	{
-		$this->render('view',array(
-			'model'=>$this->loadModel($id),
-		));
+		if(Yii::app()->user->checkAccess('readTvobject'))
+		{
+			$this->render('view',array(
+				'model'=>$this->loadModel($id),
+			));
+		}
+		else
+		{
+			$this->redirect(array('site/login'));
+		}
 	}
 
 	/**
@@ -62,21 +42,28 @@ class TvobjectController extends Controller
 	 */
 	public function actionCreate()
 	{
-		$model=new Tvobject;
-
-		// Uncomment the following line if AJAX validation is needed
-		// $this->performAjaxValidation($model);
-
-		if(isset($_POST['Tvobject']))
+		if(Yii::app()->user->checkAccess('createTvobject'))
 		{
-			$model->attributes=$_POST['Tvobject'];
-			if($model->save())
-				$this->redirect(array('view','id'=>$model->id));
+			$model=new Tvobject;
+	
+			// Uncomment the following line if AJAX validation is needed
+			// $this->performAjaxValidation($model);
+	
+			if(isset($_POST['Tvobject']))
+			{
+				$model->attributes=$_POST['Tvobject'];
+				if($model->save())
+					$this->redirect(array('view','id'=>$model->id));
+			}
+	
+			$this->render('create',array(
+				'model'=>$model,
+			));
 		}
-
-		$this->render('create',array(
-			'model'=>$model,
-		));
+		else
+		{
+			$this->redirect(array('site/login'));
+		}
 	}
 
 	/**
@@ -86,21 +73,28 @@ class TvobjectController extends Controller
 	 */
 	public function actionUpdate($id)
 	{
-		$model=$this->loadModel($id);
-
-		// Uncomment the following line if AJAX validation is needed
-		// $this->performAjaxValidation($model);
-
-		if(isset($_POST['Tvobject']))
+		if(Yii::app()->user->checkAccess('updateTvobject'))
 		{
-			$model->attributes=$_POST['Tvobject'];
-			if($model->save())
-				$this->redirect(array('view','id'=>$model->id));
+			$model=$this->loadModel($id);
+	
+			// Uncomment the following line if AJAX validation is needed
+			// $this->performAjaxValidation($model);
+	
+			if(isset($_POST['Tvobject']))
+			{
+				$model->attributes=$_POST['Tvobject'];
+				if($model->save())
+					$this->redirect(array('view','id'=>$model->id));
+			}
+	
+			$this->render('update',array(
+				'model'=>$model,
+			));
 		}
-
-		$this->render('update',array(
-			'model'=>$model,
-		));
+		else
+		{
+			$this->redirect(array('site/login'));
+		}
 	}
 
 	/**
@@ -110,11 +104,18 @@ class TvobjectController extends Controller
 	 */
 	public function actionDelete($id)
 	{
-		$this->loadModel($id)->delete();
-
-		// if AJAX request (triggered by deletion via admin grid view), we should not redirect the browser
-		if(!isset($_GET['ajax']))
-			$this->redirect(isset($_POST['returnUrl']) ? $_POST['returnUrl'] : array('admin'));
+		if(Yii::app()->user->checkAccess('deleteTvobject'))
+		{
+			$this->loadModel($id)->delete();
+	
+			// if AJAX request (triggered by deletion via admin grid view), we should not redirect the browser
+			if(!isset($_GET['ajax']))
+				$this->redirect(isset($_POST['returnUrl']) ? $_POST['returnUrl'] : array('admin'));
+		}
+		else
+		{
+			$this->redirect(array('site/login'));
+		}	
 	}
 
 	/**
@@ -122,10 +123,17 @@ class TvobjectController extends Controller
 	 */
 	public function actionIndex()
 	{
-		$dataProvider=new CActiveDataProvider('Tvobject');
-		$this->render('index',array(
-			'dataProvider'=>$dataProvider,
-		));
+		if(Yii::app()->user->checkAccess('readTvobject'))
+		{
+			$dataProvider=new CActiveDataProvider('Tvobject');
+			$this->render('index',array(
+				'dataProvider'=>$dataProvider,
+			));
+		}
+		else
+		{
+			$this->redirect(array('site/login'));
+		}
 	}
 
 	/**
@@ -133,14 +141,21 @@ class TvobjectController extends Controller
 	 */
 	public function actionAdmin()
 	{
-		$model=new Tvobject('search');
-		$model->unsetAttributes();  // clear any default values
-		if(isset($_GET['Tvobject']))
-			$model->attributes=$_GET['Tvobject'];
-
-		$this->render('admin',array(
-			'model'=>$model,
-		));
+		if(Yii::app()->user->checkAccess('updateTvobject'))
+		{
+			$model=new Tvobject('search');
+			$model->unsetAttributes();  // clear any default values
+			if(isset($_GET['Tvobject']))
+				$model->attributes=$_GET['Tvobject'];
+	
+			$this->render('admin',array(
+				'model'=>$model,
+			));
+		}
+		else
+		{
+			$this->redirect(array('site/login'));
+		}
 	}
 
 	/**
